requirements.txt
absl-py==1.0.0
addict==2.4.0
albumentations==0.4.6
backcall==0.2.0
basicsr==1.3.4.9
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
chainer==7.8.1
charset-normalizer==2.0.9
click==8.0.4
colour==0.1.5
cupy==9.6.0
cycler==0.11.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.9
editdistance==0.6.0
einops==0.4.1
fastrlock==0.8
fasttext==0.9.2
filelock==3.4.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grad-cam==1.3.9
grpcio==1.42.0
h5py==3.1.0
# Editable install with no version control (hat==0.1.0)
-e /home/zhaoyinfa/code/HAT-main
idna==3.3
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.2
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.0
kiwisolver==1.3.1
lmdb==1.3.0
lpips==0.1.4
Markdown==3.3.6
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.8
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
oauthlib==3.1.1
olefile==0.46
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
pandas==1.1.5
parso==0.7.1
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625649052827/work
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.1
psutil==5.9.2
ptflops==0.6.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-msssim==0.2.1
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentry-sdk==1.9.10
setproctitle==1.2.3
sewar==0.4.4
Shapely==1.8.2
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
statsmodels==0.12.2
tb-nightly==2.10.0a20220704
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2022.4.8
timm==0.5.4
torch==1.10.0
torchaudio==0.10.0
torchsummary @ file:///data/zhaoyinfa/torchsummary-1.5.1-py3-none-any.whl
torchvision==0.11.1
tqdm==4.62.3
traitlets==4.3.3
ttach==0.0.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.12
wandb==0.13.4
wcwidth==0.2.5
Werkzeug==2.0.2
yapf==0.32.0
zipp==3.6.0
